% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_dirs}
\alias{create_dirs}
\title{Create standard project directories}
\source{
Borrowed from my old startR package at \url{https://github.com/emlab-ucsb/startR/blob/master/R/create_local_dirs.R}
}
\usage{
create_dirs(other_dirs = NULL)
}
\arguments{
\item{other_dirs}{If needed, a character vector containing the names for extra directories.}
}
\description{
Creates a set of standard directories in the repository: scripts, data (and subdirectories for raw, processed, and output) results (and subdirectories for img and tab)
}
\examples{

\dontrun{
create_dirs() # This creates default directories

create_dirs(other_dirs = "docs") # Creates all default directories and a "docs" directory
}

}
