% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fishing}
\alias{data_fishing}
\title{Fisheries production and revenues
Information in fisheries production and revenues for 41 economic units
targeting 3 species in the Baja California Peninsula between 2000 and 2021}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1051 rows and 5 columns.
}
\source{
Data are originally reported by CONAPESCA (Mexico's fishery management agency)
and made available online at: \url{https://conapesca.gob.mx/wb/cona/avisos_arribo_cosecha_produccion}
This clean version comes from Villaseñor-Derbez, J.C., Arafeh-Dalmau, N. & Micheli, F. Past and future
impacts of marine heatwaves on small-scale fisheries in Baja California, Mexico.
Commun Earth Environ 5, 623 (2024). \url{https://doi.org/10.1038/s43247-024-01696-x}
}
\usage{
data_fishing
}
\description{
A data.frame with 1051 rows and 5 columns:
\describe{
   \item{eu_rnpa}{Character - A 10-digit unique identifier for each economic unit}
   \item{year}{Numeric - The year of the fishing seson}
   \item{fishery}{Character - Specifies the fishery (lobster, sea_cucumber, or urchin)}
   \item{catch}{Numeric - Fisherie sproduction, in Metric tons}
   \item{revenue}{Numeric - Revenue generated from selling the catch, in thousands of MXN}
}
}
\keyword{datasets}
