% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fishing_effort}
\alias{data_fishing_effort}
\title{Apparent fishing effort in the Gulf of Mexico}
\format{
## `data_fishing_effort`
A data frame with 14,038 rows and 6 columns:
\describe{
  \item{year}{Numeric - Year of the observation}
  \item{month}{Numeric - Month of the observation}
  \item{lat}{Numeric - Latitude in decimal degrees indicating the center of the pixel where activity took place}
  \item{lon}{Numeric - Longitude in decimal degrees indicating the center of the pixel where activity took place}
  \item{effort_vessels}{Numeric - Fishing effort measured in number of fishing vessels infered to be fishing at that location}
  \item{temperature_C}{Numeric - Fishing effort measured in apparent fishing hours at that location}
}
}
\source{
Data were generated using the `gfwr` package
\url{https://github.com/GlobalFishingWatch/gfwr}. See also Kroodsma
et al. "Tracking the global footprint of fisheries." Science 359.6378 (2018):
904-908.
}
\usage{
data_fishing_effort
}
\description{
Monthly AIS-derived dishing effort in the Gulf of Mexico during 2024
}
\keyword{datasets}
