% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_geartypes}
\alias{data_geartypes}
\title{Fishing effort by gear type and vessel}
\format{
## `data_geartypes`
A data frame with 840 rows and 3 columns:
\describe{
  \item{vessel_id}{Character - Unique vessel identifier}
  \item{geartype}{Character - Type of gear}
  \item{effort_hours}{Numeric - Total fishing effort, measured in hours}
}
}
\source{
Data were generated using the `gfwr` package
\url{https://github.com/GlobalFishingWatch/gfwr}. See also Kroodsma
et al. "Tracking the global footprint of fisheries." Science 359.6378 (2018):
904-908.
}
\usage{
data_geartypes
}
\description{
AIS-derived dishing effort for vessels in the Gulf of Mexico during 2024
}
\keyword{datasets}
