% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_hurricanes}
\alias{data_hurricanes}
\title{Hurricane tracks as an sf object}
\format{
## `data_hurricanes`
An sf object with 51 rows and 4 columns:
\describe{
  \item{season}{Numeric - Indicates the year in which the storm was named}
  \item{name}{Character - Storm name given by the agency}
  \item{max_sshs}{Character - Maximum value attained by the stormed, using the Saffir–Simpson hurricane scale. The values are defined as follows:
  \describe{
     \item{-5}{Unknown}
     \item{-4}{Post-tropical}
     \item{-3}{Miscellaneous disturbances}
     \item{-2}{Subtropical}
     Tropical systems classified based on wind speeds
     \item{-1}{Tropical depression (W<34)}
     \item{0}{Tropical storm (34<W<64)}
     \item{1}{Category 1 (64<=W<83)}
     \item{2}{Category 2 (83<=W<96)}
     \item{3}{Category 3 (96<=W<113)}
     \item{4}{Category 4 (113<=W<137)}
     \item{5}{Category 5 (W >= 137)}
  }
  }
  \item{geometry}{sfc -  A simple feature column containing LINESTRINGS for storm tracks}
}
}
\source{
Data come from: \url{https://www.ncei.noaa.gov/products/international-best-track-archive}
}
\usage{
data_hurricanes
}
\description{
Observed hurricane tracks for 64 storms in the North Atlantic (2022-2024).
}
\keyword{datasets}
