% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_lionfish}
\alias{data_lionfish}
\title{Lionfish biometry}
\format{
## `data_lionfish`
A data frame with 109 rows and 8 columns:
\describe{
  \item{id}{Character - A unique alpha-numeric code identoifying each organism}
  \item{site}{Character - Name of the dive site from which the organism was collected}
  \item{lat}{Numeric - Latitude in decimal degrees}
  \item{lon}{Numeric - Longitude in decimal degrees}
  \item{total_length_mm}{Numeric - Total length of the organism, measured in milimiters}
  \item{total_weight_gr}{Numeric - Total weight (also known as wet weight) of the organism, measured in milimiters}
  \item{size_class}{Character - Size class specifying whether the fish was small (< 100 mm), medium (100 - 200 mm) or large (> 300 mm)}
  \item{depth_m}{Numeric - Depth at which the organism was collected, in meters}
  \item{temperature_C}{Numeric - Water temperature during time of collection, in °C}
}
}
\source{
Villaseñor-Derbez, J.C. & Herrera-Pérez, R. Brief description of prey
selectivity and ontogenetic changes in the diet of the invasive lionfish
_Pterois volitans_ (Actinopterygii, Scorpaenidae) in the Mexican Caribbean.
Pan-American Journal of Aquatic Sciences. (2014), 9(2):131-135.
}
\usage{
data_lionfish
}
\description{
Biometric measurements for 109 lionfish (_Pterois volitans_) captured off
Puerto Aventuras (Mexico). Data also include spatial and environmental variables.
}
\keyword{datasets}
