% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mhw_events}
\alias{data_mhw_events}
\title{Marine Heatwaves in TURF-managed fisheries}
\format{
## `data_mhw_events`
A data.frame with 82 rows and 5 columns:
\describe{
   \item{date_peak}{Date - The date in which the event was at its maximum}
   \item{intensity_mean}{Numeric - Mean MHW intensity, in °C}
   \item{intensity_max}{Numeric - Max MHW intensity, in °C}
   \item{intensity_cumulative}{Numeric - Cumulative MHW intensity, in °C days}
   }
}
\source{
Villaseñor-Derbez, J.C., Arafeh-Dalmau, N. & Micheli, F. Past and future
impacts of marine heatwaves on small-scale fisheries in Baja California, Mexico.
Commun Earth Environ 5, 623 (2024). \url{https://doi.org/10.1038/s43247-024-01696-x}
}
\usage{
data_mhw_events
}
\description{
Marine heatwave events detected for one location
}
\keyword{datasets}
