% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mhw_ts}
\alias{data_mhw_ts}
\title{Marine Heatwaves in TURF-managed fisheries}
\format{
## `data_mhw_ts`
A data.frame with 1431 rows and 4 columns:
\describe{
   \item{date}{Date - The date}
   \item{temp}{Numeric - Sea surface temperature, in °C}
   \item{seas}{Numeric - Seasonality, in °C}
   \item{thresh}{Numeric - Threshold to detect as marine heatwave, in °C}
   }
}
\source{
Villaseñor-Derbez, J.C., Arafeh-Dalmau, N. & Micheli, F. Past and future
impacts of marine heatwaves on small-scale fisheries in Baja California, Mexico.
Commun Earth Environ 5, 623 (2024). \url{https://doi.org/10.1038/s43247-024-01696-x}
}
\usage{
data_mhw_ts
}
\description{
Daily time series of sea surface temperature, seasonality, and marine heatwave thresholds for a single site
}
\keyword{datasets}
