% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_milton}
\alias{data_milton}
\title{Hurricane tracks for Milton (2024)}
\format{
## `data_milton`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{name}{Character - Storm name given by the agency}
  \item{iso_time}{POSIXct - SO Time provided in Universal Time Coordinates (UTC). Format is YYYY-MM-DD HH:mm:ss Most points are provided at 6 hour intervals.}
  \item{lat}{Numeric - Latitude in decimal degrees}
  \item{lon}{Numeric - Longitude in decimal degrees}
  \item{wind_speed}{Numeric - Maximum sustained wind speed in knots: 0 - 300 kts. (1-min. mean)}
  \item{pressure}{Numeric - Minimum sea level pressure, 850 - 1050 mb.}
  \item{sshs}{Numeric - Saffir–Simpson hurricane scale. The values are defined as follows:
  -5 = Unknown
  -4 = Post-tropical
  -3 = Miscellaneous disturbances
  -2 = Subtropical
  Tropical systems classified based on wind speeds
  -1 = Tropical depression (W<34)
  0 = Tropical storm (34<W<64)
  1 = Category 1 (64<=W<83)
  2 = Category 2 (83<=W<96)
  3 = Category 3 (96<=W<113)
  4 = Category 4 (113<=W<137)
  5 = Category 5 (W >= 137)
  }
}
}
\source{
\url{https://www.ncei.noaa.gov/products/international-best-track-archive}
}
\usage{
data_milton
}
\description{
Observed hurricane track (lat, lon, timestamp) for center of storm Milton during 2024
}
\keyword{datasets}
