% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_milton_sf}
\alias{data_milton_sf}
\title{Hurricane tracks for Milton (2024) as an sf object}
\format{
## `data_milton_sf`
A data frame with 62 rows and 7 columns:
\describe{
  \item{name}{Character - Storm name given by the agency}
  \item{iso_time}{POSIXct - SO Time provided in Universal Time Coordinates (UTC). Format is YYYY-MM-DD HH:mm:ss Most points are provided at 6 hour intervals.}
  \item{wind_speed}{Numeric - Maximum sustained wind speed in knots: 0 - 300 kts. (1-min. mean)}
  \item{pressure}{Numeric - Minimum sea level pressure, 850 - 1050 mb.}
  \item{geometrt}{sfc -  A simple fieature column containing points for the location of the storm center}
  \item{sshs}{Character - Saffir–Simpson hurricane scale. The values are defined as follows:
  \describe{
     \item{-5}{Unknown}
     \item{-4}{Post-tropical}
     \item{-3}{Miscellaneous disturbances}
     \item{-2}{Subtropical}
     Tropical systems classified based on wind speeds
     \item{-1}{Tropical depression (W<34)}
     \item{0}{Tropical storm (34<W<64)}
     \item{1}{Category 1 (64<=W<83)}
     \item{2}{Category 2 (83<=W<96)}
     \item{3}{Category 3 (96<=W<113)}
     \item{4}{Category 4 (113<=W<137)}
     \item{5}{Category 5 (W >= 137)}
  }
  }
}
}
\source{
\url{https://www.ncei.noaa.gov/products/international-best-track-archive}
}
\usage{
data_milton_sf
}
\description{
Observed hurricane track for center of storm Milton during 2024. This is a spatial version of data_milton
}
\keyword{datasets}
