% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_MPA}
\alias{data_MPA}
\title{SIMULATED biomass inside and outside a Marine Protected Area}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 5 columns.
}
\source{
This data are \bold{SIMULATED}. For instructional purposes only.
}
\usage{
data_MPA
}
\description{
A data.frame with 100 rows and 5 columns:
\describe{
   \item{time}{Numeric - Number of year until / since the the intervention takes place}
   \item{id}{Character - Unique site identifier (letters a:j)}
   \item{protected}{Numeric - A value of 1, indicates the observation comes from inside the MPA }
   \item{after}{Numeric - A value of 1 indicates the observation comes from the period after the intervention took place}
   \item{biomass}{Numeric - \bold{SIMULATED} biomass data, in Kg}
}
}
\details{
Panel data with \bold{SIMULATED} biomass data from inside and
outside a Marine Protected Area. The average treatment effect is 2, but noise
has been introuced to the data.
}
\keyword{datasets}
