% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_PIPA}
\alias{data_PIPA}
\title{CPUE near and far the Phoenix Islands Protected Area}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 663 rows and 6 columns.
}
\usage{
data_PIPA
}
\description{
A data.frame with 663 rows and 6 columns:
\describe{
   \item{year}{Numeric - Number of year until / since the the intervention takes place}
   \item{lat}{Numeric - Latitude indicating centroid of the grid cell}
   \item{lon}{Numeric - Longitude indicating centroid of the grid cell}
   \item{period}{Character - indicates whether the observation comes from near (within 100 nautical miles) or far (between 100 and 200 nautical miles) from the MPA }
   \item{distance}{Characger - indicates whether the observation comes before or after the intervention took place}
   \item{cpue}{Numeric - Catch-per-unit-effort meassured in metric tons per fishing set}
}
}
\details{
Panel data with CPUE data from near and far from the boundaries of the
Phoenix Islands Protected Area.
}
\keyword{datasets}
