% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sst}
\alias{data_sst}
\title{Sea Surface Temperature}
\format{
## `data_sst`
A data frame with 2,922 rows and two columns:
\describe{
   \item{date}{Date - Date of the observation, in yyyy-mm-dd format}
   \item{temperature_C}{Numeric - Mean sea surface temperatrure, measured in °C}
}
}
\source{
Raw data come from NOAA AVHRR Optimum Interpolation v2.1 - SST
\url{https://coastwatch.pfeg.noaa.gov/erddap/griddap/ncdcOisst21Agg_LonPM180.html}.
The script to generate the data is \url{https://github.com/jcvdav/EVR628tools/blob/main/data-raw/data_sst.R}
}
\usage{
data_sst
}
\description{
Mean daily Sea Surface Temperature (°C) derived from 0.5° resolution imagery
within bounding box (lon_min = -81, lon_max = -79.75, lat_min = 25.25,
lat_max = 25.75).
}
\examples{
library(EVR628tools)
library(ggplot2)
# Look at the first 10 rows
head(data_sst, 10)
# Make a time series plot
ggplot(data = EVR628tools::data_sst,
       mapping = aes(x = date, y = temperature_C)) +
       geom_line()
}
\keyword{datasets}
