% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sst_anom}
\alias{data_sst_anom}
\title{Sea Surface Temperature Anomaly}
\format{
## `data_sst_anom`
A data.frame with 9 rowns and two columns:
\describe{
  \item{lat}{Numeric - Latitude in decimal degrees indicating the center of the pixel}
  \item{lon}{Numeric - Longitude in decimal degrees indicating the center of the pixel}
  \item{temperature_anomaly_C}{Numeric - Sea surface temperatrure anomaly relative to daily 2000-2023 values, measured in °C}
}
}
\source{
Raw data come from NOAA AVHRR Optimum Interpolation v2.1 - SST
\url{https://coastwatch.pfeg.noaa.gov/erddap/griddap/ncdcOisst21Agg_LonPM180.html}.
The script to generate the data is \url{https://github.com/jcvdav/EVR628tools/blob/main/data-raw/data_sst.R}
}
\usage{
data_sst_anom
}
\description{
Sea Surface Temperature (°C) derived from 0.5° resolution imagery
within bounding box (lon_min = -81, lon_max = -79.75, lat_min = 25.25,
lat_max = 25.75) for 2024, relative to 2000-2023 values.
}
\examples{
# Look at the first 10 rows
head(EVR628tools::data_sst_anom, 10)
}
\keyword{datasets}
