% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_IPCC.R
\name{palette_IPCC}
\alias{palette_IPCC}
\title{IPCC color palettes}
\source{
Page 8 of the 2019 ICPP Style guide: \url{https://www.ipcc.ch/site/assets/uploads/2019/04/IPCC-visual-style-guide.pdf}

Mark Harrower & Cynthia A. Brewer (2003) ColorBrewer.org: An Online Tool for Selecting Colour Schemes forMaps, The Cartographic Journal, 40:1, 27-37, DOI: 10.1179/000870403235002042
}
\usage{
palette_IPCC(var = "temp", type = "div")
}
\arguments{
\item{var}{Character string indicating the type of variable. Options are "temp", "prec", "wind", "cryo"}

\item{type}{Character string specifying if the user wants divergent ("div") or sqeuential "seq" palettes}
}
\value{
A character vector containing hex color codes
}
\description{
Provides standard divergent and sequential IPCC color palettes for temperature, precipitation, wind and cryosphere
}
