% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_IPCC_ssp.R
\name{palette_IPCC_ssp}
\alias{palette_IPCC_ssp}
\title{Color palette for the Shared Socio-economic Pathways (SSPs)}
\source{
Page 9 form the IPCC style gyide: \url{https://www.ipcc.ch/site/assets/uploads/2022/09/IPCC_AR6_WGI_VisualStyleGuide_2022.pdf}
}
\usage{
palette_IPCC_ssp(
  ssp = c("SSP5-8.5", "SSP3-7.0", "SSP2-4.5", "SSP1-2.6", "SSP1-1.9")
)
}
\arguments{
\item{ssp}{A character vector specifying which SSPs are required. Options are `"SSP5-8.5"`, `"SSP3-7.0"`, `"SSP2-4.5"`, `"SSP1-2.6"` and `"SSP1-1.9"`}
}
\value{
A NAMED vector containing pairs for SSP names and colors.
}
\description{
Representative Concentration Pathways (RCPs) and Socio-economic Pathways (SSPs), as indicated by the IPCC AR6 WGI style guide
}
