% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_IPCC_temp.R
\name{palette_IPCC_temp}
\alias{palette_IPCC_temp}
\title{IPCC color palette for Sea Surface Temperature
#' @description
Default colors for Sea Surface Temperature, as indicated by the IPCC AR6 WGI style guide, when using a divergent color palette}
\source{
Page 8 of the 2019 ICPP Style guide: \url{https://www.ipcc.ch/site/assets/uploads/2019/04/IPCC-visual-style-guide.pdf}

Mark Harrower & Cynthia A. Brewer (2003) ColorBrewer.org: An Online Tool for Selecting Colour Schemes forMaps, The Cartographic Journal, 40:1, 27-37, DOI: 10.1179/000870403235002042
}
\usage{
palette_IPCC_temp()
}
\value{
A character vector containing 11 colors,
}
\description{
IPCC color palette for Sea Surface Temperature
#' @description
Default colors for Sea Surface Temperature, as indicated by the IPCC AR6 WGI style guide, when using a divergent color palette
}
\examples{
library(EVR628tools)
palette_IPCC_temp()
}
