% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_UM.R
\name{palette_UM}
\alias{palette_UM}
\title{UM Color palette}
\source{
See the University of Miami's Visual Identity Cues for Web: \url{https://webcomm.miami.edu/resources/identity/color/index.html}
}
\usage{
palette_UM(n, require_WCAG = FALSE)
}
\arguments{
\item{n}{An integer. Number of colors to return, from a maximum of 10.}

\item{require_WCAG}{A boolean specifying whether colors returned should comply with Web Content Accessibility Guidelines (WCAG) 2 Level AA Conformance (See \url{https://www.w3.org/WAI/WCAG2AA-Conformance}). Defaults to FALSE.}
}
\value{
A string of HEX color values
}
\description{
A discrete color palette with up to 10 colors based on the Unviersity of Miami's color identity guidelines.
}
\examples{

library(EVR628tools)
# Ask for 5 colors
palette_UM(5)

}
