% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sst}
\alias{data_sst}
\title{Sea Surface Temperature}
\format{
## `data_sst`
\describe{
   \item{date}{Date - Date of the observation, in yyyy-mm-dd format}
   \item{temperature_C}{Numeric - Mean sea surface temperatrure, measured in °C}
}
}
\usage{
data_sst
}
\description{
Mean daily Sea Surface Temperature (°C) near Miami, FL
}
\examples{
library(EVR628tools)
# Look at the first 10 rows
head(data_sst, 10)
# Make a time series plot
ggplot(data = data_sst,
       mapping = aes(x = date, y = temperature_C)) +
       geom_line()
}
\keyword{datasets}
